Consolidate and comment similar types.
