--- Sources/muterCore/Configuration/SourceFileInfo.swift
@@ -1,9 +1,0 @@
-/// The full path to a file on disk, and its complete contents as a string.
-struct SourceFileInfo {
-
-    /// The path to the file on disk.
-    let file: String
-
-    /// The full contents of the file.
-    let source: String
-}
